+++ CHANGELOG.md
@@ -18,11 +18,15 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 * Load model state dict into a new model with modified `RPUConfig`. (\#276)
 * Visualization for noise models for analog inference hardware simulation. (\#278)
+* State indipendent inference noise model. (\# 284)
+* Transfer LR parameter for ``MixedPrecisionCompound``. (\#283)
 
 ### Fixed
 
 * Removed GPU warning during destruction when using multiple GPUs. (\#277)
 * Mixed precision ... .(\#292)
+* Fix issue in transfer counter for mixed precision in case of GPU. (\#283)
+
 
 ### Changed